Rebase, merge and update tests
