+++ client/web/src/integration/graphQlResponseHelpers.ts
@@ -1,4 +1,5 @@
 import { encodeURIPathComponent } from '@sourcegraph/common'
+import { JsonDocument } from '@sourcegraph/shared/src/codeintel/scip'
 import { TreeEntriesResult } from '@sourcegraph/shared/src/graphql-operations'
 
 import {
@@ -31,7 +32,8 @@ export const createTreeEntriesResult = (url: string, toplevelFiles: string[]): T
 
 export const createBlobContentResult = (
     content: string,
-    html: string = `<div style="color:red">${content}<div>`
+    html: string = `<div style="color:red">${content}<div>`,
+    lsif?: JsonDocument
 ): BlobResult => ({
     repository: {
         commit: {
@@ -41,7 +43,7 @@ export const createBlobContentResult = (
                 highlight: {
                     aborted: false,
                     html,
-                    lsif: '',
+                    lsif: lsif ? JSON.stringify(lsif) : '',
                 },
             },
         },
